worker/requirements.txt
pika
redis
python-dotenv
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-qdrant
qdrant-client
pypdf